chore(query builder): rename builder
